API-Kafka-mySQL/requirement.txt
mysql-connector-python==8.0.25
protobuf==3.17.1
six==1.16.0
